Deleted scripts from device/bin/silicon
These scripts are out of date and its unclear if anyone is
actually interested in using them. So I am removing them to remove the
expectation that they will be maintained. If there is a need for one of
these scripts we can re-add it with a specific agreement for who is
responsible for keeping it up to date.