Fix l1 membar log (#218)
Related tt_metal CI issue:
https://github.com/tenstorrent/tt-metal/issues/14225
This was accidentally checked in #64 

This filled up logs on CI machines that tt_metal uses and started
crashing them. This is non-deterministic behavior and also some later
jobs might fail due to this issue, not exactly the job responsible for
it -> which both made it hard to find the cause of the issue.

Hopefully this is the only issue, if there is an additional issue with
UMD being stuck we will investigate further.