Add crude benchmark
This is bad because the outer benchmark (in the unit test) should not
measure the cout time in the inner benchmark (in pci_device.cpp).  I
"solved" this by commenting them out back and forth.  So that is problem
1.

Problem 2 is that this test does not capture the nuance of the IO code
path.  There can be static vs dynamic TLBs (this hits dynamic path).
There can be write-combined vs uncached memory mappings (this is going
through WC).

It does however demonstrate that there is a drastic amount of overhead
imposed by the UMD implementation.